+++ src/pages/point-detail/index.tsx
@@ -6,6 +6,7 @@ import { numberFormat, removeNumberFormat } from './numberFormat';
 import { ExclamationCircleOutlined } from '@ant-design/icons';
 import { TextBox } from '@components/text-box';
 import { colors } from '@/constants/colors';
+import { InfoContainer } from './info-container';
 const MINIMUM_PRICE = 10000;
 const MAXIMUM_PRICE = 10000000;
 
@@ -26,6 +27,10 @@ export const PointDetail = () => {
       setFormattedValue('');
     }
 
+    priceComparator(inputValue);
+  };
+
+  const priceComparator = (inputValue: string) => {
     if (
       parseInt(removeNumberFormat(inputValue)) < MINIMUM_PRICE ||
       inputValue === ''
@@ -117,31 +122,7 @@ export const PointDetail = () => {
             </PointButtonWrap>
 
             <PriceWrap>
-              {isInfoBoxState && (
-                <InfoPriceContainer direction="vertical" size={0}>
-                  <div>
-                    <TextBox
-                      typography="body3"
-                      color={'black900'}
-                      fontWeight={'400'}
-                    >
-                      충전하시는 포인트에 따른 결제 금액입니다.
-                    </TextBox>
-                  </div>
-
-                  <div>
-                    <TextBox
-                      typography="body3"
-                      color={'black900'}
-                      fontWeight={'400'}
-                      style={{ position: 'relative' }}
-                    >
-                      *포인트 환산 기준: 결제 금액 = 포인트
-                    </TextBox>
-                    <TailStyle></TailStyle>
-                  </div>
-                </InfoPriceContainer>
-              )}
+              {isInfoBoxState && <InfoContainer />}
 
               <InfoButton
                 onClick={() => {
@@ -278,34 +259,7 @@ const AgreementWrap = styled(Space)`
   justify-content: center;
   text-align: center;
 `;
-const InfoPriceContainer = styled(Space)`
-  position: absolute;
-  bottom: 196px;
-  right: 54px;
 
-  width: 266px;
-
-  border-radius: 2px;
-
-  padding: 6px 8px;
-
-  box-shadow: 0px 2px 8px rgba(0, 0, 0, 0.15);
-
-  text-align: center;
-
-  gap: 0px;
-`;
-const TailStyle = styled('div')`
-  width: 8px;
-  height: 8px;
-
-  position: absolute;
-  bottom: -4px;
-  left: 50%;
-
-  transform: rotate(45deg);
-  background-color: white;
-`;
 const InfoButton = styled('button')`
   background: rgba(0, 0, 0, 0);
   border: none;
